[2025.05.21] - 김건우 1. trading 매도, 매수 시 보유 자산 update 로직 추가

+++ src/main/java/com/kwfw/findiary/model/BankAccountDto.java
@@ -14,8 +14,10 @@ public class BankAccountDto {
     private String create_date;
     private String modify_date;
 
-    private int account_total_property;
+    private float account_total_property;
     private String currency;
+
+
 
     private List<DiaryDto> diaryList;
 }
